backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.26.4
scipy==1.11.2
Pillow==10.0.0
svgwrite==1.4.3